backend/requirements.txt
langchain==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
python-dotenv==1.0.1
openai==1.30.1
# langchain-chroma==0.1.0
langchain-community==0.2.0
chromadb==0.4.24
pypdf==4.2.0
python-docx==1.1.2
docx2txt==0.8